requirements.txt
black
faiss-cpu
fastapi
matplotlib
mongomock
numpy
openai
pandas
pydantic
pydantic-settings
pytube
pyyaml
pymongo
pymupdf
pytest
python-dotenv
python-pptx
requests
scikit-learn
seaborn
sentence-transformers
tabulate
tensorflow_hub
tensorflow_text
transformers
torch
tqdm
uvicorn[standard]
yt_dlp
wandb

